ts paths and file rename for HomeSideBar Component
