+++ apps/next/pages/distributions.tsx
@@ -1,6 +1,6 @@
 import { DistributionsScreen } from 'app/features/distributions/screen'
 import Head from 'next/head'
-import { HomeSideBarWrapper } from 'app/components/sidebar'
+import { HomeSideBarWrapper } from 'app/components/sidebar/HomeSideBar'
 import { NextPageWithLayout } from './_app'
 import { userProtectedGetSSP } from 'utils/userProtected'
 import { useRouter } from 'next/router'
